+++ bindings/python/CHANGELOG.md
@@ -30,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Made `TaggedDataPayload::data` optional;
+- Made `TaggedDataPayload::{tag, data}` optional;
 
 ## 1.1.2 - 2023-12-01