+++ bindings/python/CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `NodeInfoProtocol.belowMaxDepth`;
 - `{NodeInfoProtocol, RentStructure}::as_dict()`;
 
+### Changed
+
+- Made the type of `TaggedDataPayload` required as first field in `__init__()`;
+
 ### Fixed
 
 - Made `TaggedDataPayload::{tag, data}` optional;